Put console code in separate file
